+++ apps/expo-go/android/expoview/src/main/java/host/exp/exponent/experience/splashscreen/legacy/SplashScreenView.kt
@@ -29,6 +29,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
+import com.bumptech.glide.integration.compose.placeholder
+import host.exp.expoview.R
 
 // this needs to stay for versioning to work
 
@@ -100,6 +102,7 @@ fun SplashScreenImage(
     model = imageUrl,
     contentDescription = "Splash Screen Image",
     contentScale = resizeMode.toContentScale(),
+    loading = placeholder(R.drawable.project_default_icon),
     modifier = Modifier
       .width(width)
       .height(height)