+++ apps/router-e2e/__e2e__/headless/components/TabButton.tsx
@@ -1,7 +1,7 @@
 import FontAwesome from '@expo/vector-icons/FontAwesome';
 import { TabTriggerSlotProps } from 'expo-router/ui';
-import { ComponentProps, Ref, forwardRef } from 'react';
-import { Text, StyleSheet, Pressable, View } from 'react-native';
+import { ComponentProps } from 'react';
+import { Text, StyleSheet, Pressable } from 'react-native';
 
 type Icon = ComponentProps<typeof FontAwesome>['name'];
 
@@ -9,21 +9,16 @@ export type TabButtonProps = TabTriggerSlotProps & {
   icon?: Icon;
 };
 
-export const TabButton = forwardRef(
-  ({ icon, children, isFocused, ...props }: TabButtonProps, ref: Ref<View>) => {
-    return (
-      <Pressable
-        ref={ref}
-        {...props}
-        style={[styles.button, isFocused ? styles.focusedBackground : undefined]}>
-        <FontAwesome name={icon} />
-        <Text style={[styles.tabTriggerText, isFocused ? styles.focusedText : undefined]}>
-          {children}
-        </Text>
-      </Pressable>
-    );
-  }
-);
+export function TabButton({ icon, children, isFocused, ...props }: TabButtonProps) {
+  return (
+    <Pressable {...props} style={[styles.button, isFocused ? styles.focusedBackground : undefined]}>
+      <FontAwesome name={icon} />
+      <Text style={[styles.tabTriggerText, isFocused ? styles.focusedText : undefined]}>
+        {children}
+      </Text>
+    </Pressable>
+  );
+}
 
 const styles = StyleSheet.create({
   root: {