+++ packages/expo-notifications/src/NotificationsHandler.ts
@@ -77,7 +77,8 @@ let handleTimeoutSubscription: EventSubscription | null = null;
  *
  * Notifications.setNotificationHandler({
  *   handleNotification: async () => ({
- *     shouldShowAlert: true,
+ *     shouldShowBanner: true,
+ *     shouldShowList: true,
  *     shouldPlaySound: false,
  *     shouldSetBadge: false,
  *   }),
@@ -110,6 +111,12 @@ export function setNotificationHandler(handler: NotificationHandler | null): voi
         try {
           const mappedNotification = mapNotification(notification);
           const behavior = await handler.handleNotification(mappedNotification);
+
+          if (behavior.shouldShowAlert) {
+            console.warn(
+              '[expo-notifications]: `shouldShowAlert` is deprecated. Specify `shouldShowBanner` and / or `shouldShowList` instead.'
+            );
+          }
           await NotificationsHandlerModule.handleNotificationAsync(id, behavior);
           handler.handleSuccess?.(id);
         } catch (error: any) {