+++ packages/expo/ios/AppDelegates/ExpoAppDelegate.swift
@@ -16,28 +16,17 @@ import ReactAppDependencyProvider
 open class ExpoAppDelegate: NSObject, ReactNativeFactoryProvider, UIApplicationDelegate {
   @objc public var reactNativeFactory: RCTReactNativeFactory?
   @objc public var reactNativeFactoryDelegate: ExpoReactNativeFactoryDelegate?
+  private let defaultModuleName = "main"
+  private let defaultInitialProps = [AnyHashable: Any]()
 
   /// The window object, used to render the UIViewControllers
   public var window: UIWindow?
 
-  /// From RCTAppDelegate
-  @objc public var moduleName: String = ""
-  @objc public var initialProps: [AnyHashable: Any]?
-
-  /// If `automaticallyLoadReactNativeWindow` is set to `true`, the React Native window will be loaded automatically.
-  public var automaticallyLoadReactNativeWindow = true
-
-  func loadReactNativeWindow(launchOptions: [UIApplication.LaunchOptionsKey: Any]?) {
-#if os(iOS) || os(tvOS)
-    window = UIWindow(frame: UIScreen.main.bounds)
-    reactNativeFactory?.startReactNative(
-      withModuleName: self.moduleName,
-      in: window,
-      launchOptions: launchOptions)
-#elseif os(macOS)
+  func loadMacOSWindow(launchOptions: [UIApplication.LaunchOptionsKey: Any]?) {
+#if os(macOS)
     if let rootView = reactNativeFactory?.rootViewFactory.view(
-      withModuleName: self.moduleName as String,
-      initialProperties: self.initialProps,
+      withModuleName: defaultModuleName,
+      initialProperties: defaultInitialProps,
       launchOptions: launchOptions
     ) {
       let frame = NSRect(x: 0, y: 0, width: 1280, height: 720)
@@ -47,7 +36,7 @@ open class ExpoAppDelegate: NSObject, ReactNativeFactoryProvider, UIApplicationD
         backing: .buffered,
         defer: false)
 
-      window.title = moduleName
+      window.title = defaultModuleName
       window.autorecalculatesKeyViewLoop = true
 
       let rootViewController = NSViewController()
@@ -87,22 +76,19 @@ open class ExpoAppDelegate: NSObject, ReactNativeFactoryProvider, UIApplicationD
       }
     }
 
-    self.moduleName = moduleName ?? ""
-    self.initialProps = initialProps
-
     let rootView: UIView
     if let factory = rootViewFactory as? ExpoReactRootViewFactory {
       // When calling `recreateRootViewWithBundleURL:` from `EXReactRootViewFactory`,
       // we don't want to loop the ReactDelegate again. Otherwise, it will be an infinite loop.
       rootView = factory.superView(
-        withModuleName: self.moduleName as String,
-        initialProperties: self.initialProps ?? [:],
+        withModuleName: moduleName ?? defaultModuleName,
+        initialProperties: initialProps,
         launchOptions: launchOptions ?? [:]
       )
     } else {
       rootView = rootViewFactory.view(
-        withModuleName: self.moduleName as String,
-        initialProperties: self.initialProps,
+        withModuleName: moduleName ?? defaultModuleName,
+        initialProperties: initialProps,
         launchOptions: launchOptions
       )
     }
@@ -137,10 +123,6 @@ open class ExpoAppDelegate: NSObject, ReactNativeFactoryProvider, UIApplicationD
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
   ) -> Bool {
-    if automaticallyLoadReactNativeWindow {
-      loadReactNativeWindow(launchOptions: launchOptions)
-    }
-
     ExpoAppDelegateSubscriberRepository.subscribers.forEach { subscriber in
       // Subscriber result is ignored as it doesn't matter if any subscriber handled the incoming URL – we always return `true` anyway.
       _ = subscriber.application?(application, didFinishLaunchingWithOptions: launchOptions)
@@ -161,10 +143,8 @@ open class ExpoAppDelegate: NSObject, ReactNativeFactoryProvider, UIApplicationD
   }
 
   open func applicationDidFinishLaunching(_ notification: Notification) {
-    if automaticallyLoadReactNativeWindow {
-      let launchOptions = notification.userInfo as? [String: Any]
-      loadReactNativeWindow(launchOptions: launchOptions)
-    }
+    let launchOptions = notification.userInfo as? [String: Any]
+    loadMacOSWindow(launchOptions: launchOptions)
 
     ExpoAppDelegateSubscriberRepository
       .subscribers